backend/requirements.txt
Flask
Werkzeug
psycopg2-binary
SQLAlchemy
pytest
requests
scikit-learn
joblib
pandas
Flask-Cors